requirements/base.txt

django==1.11.29
django-braces==1.11.0
django-environ==0.4.3
django-extensions==1.7.9
django-model-utils==3.0.0
pillow==6.2.2
psycopg2==2.7.1
pytz==2017.2
raven==6.1.0
django-solo==1.1.2
django-silk==1.0.0
django-phonenumber-field[phonenumbers]==2.0.0
django-polymorphic==2.0.2
